✨ LOCKFILE (WIP)
Add lockfile (`.floss_funding.lock`) handling to solve this problem if child process exits.
When lockfile exists, it contraindicates Poke.new.
This file is created on load, and deleted on exit.
If in DEBUG mode, check if the file exists already on load and use FlossFunding.debug_log to inform the user, noting that may be a subprocess.
It should store the PID of the process, and nothing else (like a pidfile). The lockfile should be created in the project_root directory by default, and can be overridden via ENV variable "FLOSS_FUNDING_CFG_LOCK", which must be a filename ending in .lock, a relative path from project_root, or an absolute path beginning with `File::SEPARATOR`.

development rake task that will iterate over a set of shell commands. For each iteration it will capture all output, and then parse the output, counting the number of occurrences of `FLOSS Funding Summary` within. It will print a table summary of the results for each task, with the task in column 1, and the count in column 2.